refine how we invoke tailcall optimization during compilation
